Refactor end chat functions
