+++ src/components/profile/profile.tsx
@@ -1,14 +1,13 @@
 import React, { useEffect, useState } from "react";
 import { motion } from "framer-motion";
 import { useTranslation } from "react-i18next";
-import { endChat, setIsChatOpen } from "../../slices/chat-slice";
+import { setIsChatOpen } from "../../slices/chat-slice";
 import Buerokratt from "../../static/icons/buerokratt.svg";
 import { useAppDispatch } from "../../store";
 import { getFromSessionStorage } from "../../utils/session-storage-utils";
 import styles from "./profile.module.scss";
 import useWidgetSelector from "../../hooks/use-widget-selector";
-import useChatSelector from "../../hooks/use-chat-selector";
-import { CHAT_EVENTS } from "../../constants";
+import useReloadChatEndEffect from "../../hooks/use-reload-chat-end-effect";
 
 export const Profile = (): JSX.Element => {
   const { t } = useTranslation();
@@ -16,7 +15,6 @@ export const Profile = (): JSX.Element => {
   const { widgetConfig } = useWidgetSelector();
   const [delayFinished, setDelayFinished] = useState(false);
   const newMessagesAmount = getFromSessionStorage("newMessagesAmount");
-  const { chatId } = useChatSelector();
 
   const openChat = () => {
     dispatch(setIsChatOpen(true));
@@ -38,6 +36,8 @@ export const Profile = (): JSX.Element => {
     );
   }, []);
 
+  useReloadChatEndEffect();
+  
   const getActiveProfileClass = () => {
     if (delayFinished && widgetConfig.animation === "jump")
       return styles.profile__jump;
@@ -46,40 +46,6 @@ export const Profile = (): JSX.Element => {
     if (delayFinished && widgetConfig.animation === "shockwave")
       return styles.profile__shockwave;
   };
-
-  useEffect(() => {
-    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
-      if (chatId) {
-        dispatch(
-          endChat({
-            event: CHAT_EVENTS.CLIENT_LEFT_FOR_UNKNOWN_REASONS,
-            isUpperCase: true,
-          })
-        );
-      }
-    };
-
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === "F5" || (event.ctrlKey && event.key === "r")) {
-        if (chatId) {
-          dispatch(
-            endChat({
-              event: CHAT_EVENTS.CLIENT_LEFT_FOR_UNKNOWN_REASONS,
-              isUpperCase: true,
-            })
-          );
-        }
-      }
-    };
-
-    window.addEventListener("beforeunload", handleBeforeUnload);
-    window.addEventListener("keydown", handleKeyDown);
-
-    return () => {
-      window.removeEventListener("beforeunload", handleBeforeUnload);
-      window.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [chatId]);
 
   return (
     <div className={styles.profile__wrapper}>
