+++ src/hooks/use-reload-chat-end-effect.ts
@@ -1,0 +1,39 @@
+import { useEffect } from "react";
+import { useAppDispatch } from "../store";
+import { CHAT_EVENTS } from "../constants";
+import { endChat } from "../slices/chat-slice";
+import useChatSelector from "./use-chat-selector";
+
+const useReloadChatEndEffect = () => {
+  const { chatId } = useChatSelector();
+  const dispatch = useAppDispatch();
+
+  useEffect(() => {
+    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
+      if (chatId) {
+        dispatch(
+          endChat({
+            event: CHAT_EVENTS.CLIENT_LEFT_FOR_UNKNOWN_REASONS,
+            isUpperCase: true,
+          })
+        );
+      }
+    };
+
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (event.key === "F5" || (event.ctrlKey && event.key === "r")) {
+        handleBeforeUnload(event);
+      }
+    };
+
+    window.addEventListener("beforeunload", handleBeforeUnload);
+    window.addEventListener("keydown", handleKeyDown);
+
+    return () => {
+      window.removeEventListener("beforeunload", handleBeforeUnload);
+      window.removeEventListener("keydown", handleKeyDown);
+    };
+  }, [chatId]);
+}
+
+export default useReloadChatEndEffect;
